Move GetMissingBlockIDs and Execute out of the interval package
